[WPE] Use drag implementation from GTK port https://bugs.webkit.org/show_bug.cgi?id=294144
Reviewed by NOBODY (OOPS!).

Extract common logic into *GLib.cpp files and start using them in WPE and GTK ports.
The implementation is mostly the same as was in GTK port. A few platform-specific
bits are left in corresponding *WPE.cpp and *GTK.cpp files.

* Source/WebCore/PlatformWPE.cmake:
* Source/WebCore/SourcesGTK.txt:
* Source/WebCore/SourcesWPE.txt:
* Source/WebCore/page/glib/DragControllerGLib.cpp: Renamed from Source/WebCore/page/gtk/DragControllerGtk.cpp.
This does not differ between the two platforms.

* Source/WebCore/platform/DragData.h:
* Source/WebCore/platform/DragImage.cpp:
(WebCore::createDragImageForColor):
* Source/WebCore/platform/Pasteboard.h:
* Source/WebCore/platform/glib/DragDataGLib.cpp: Renamed from Source/WebCore/platform/gtk/DragDataGtk.cpp.
This does not differ between the two platforms.

(WebCore::DragData::canSmartReplace const):
(WebCore::DragData::containsColor const):
(WebCore::DragData::containsFiles const):
(WebCore::DragData::numberOfFiles const):
(WebCore::DragData::asFilenames const):
(WebCore::DragData::containsPlainText const):
(WebCore::DragData::asPlainText const):
(WebCore::DragData::asColor const):
(WebCore::DragData::containsCompatibleContent const):
(WebCore::DragData::containsURL const):
(WebCore::DragData::asURL const):
(WebCore::DragData::shouldMatchStyleOnDrop const):
* Source/WebCore/platform/glib/PasteboardGLib.cpp: Copied from Source/WebCore/platform/gtk/PasteboardGtk.cpp.
(WebCore::pasteboardCustomDataType):
(WebCore::Pasteboard::createForCopyAndPaste):
(WebCore::Pasteboard::createForDragAndDrop):
(WebCore::Pasteboard::create):
(WebCore::Pasteboard::Pasteboard):
(WebCore::Pasteboard::selectionData const):
(WebCore::selectionDataTypeFromHTMLClipboardType):
(WebCore::Pasteboard::writeString):
(WebCore::Pasteboard::writePlainText):
(WebCore::Pasteboard::write):
(WebCore::Pasteboard::writeTrustworthyWebURLsPboardType):
(WebCore::Pasteboard::clear):
(WebCore::Pasteboard::setDragImage):
(WebCore::Pasteboard::read):
(WebCore::Pasteboard::hasData):
(WebCore::Pasteboard::typesSafeForBindings):
(WebCore::Pasteboard::typesForLegacyUnsafeBindings):
(WebCore::Pasteboard::readOrigin):
(WebCore::Pasteboard::readString):
(WebCore::Pasteboard::readStringInCustomData):
(WebCore::Pasteboard::fileContentState):
(WebCore::Pasteboard::writeMarkup):
(WebCore::Pasteboard::writeCustomData):
(WebCore::Pasteboard::changeCount const):
* Source/WebCore/platform/graphics/GraphicsContextGLImageExtractor.h:
* Source/WebCore/platform/gtk/PasteboardGtk.cpp:
Pasteboard::read only differs in the charset=utf-8 parameter from WPE, not
sure if it's important or they can be merged.

(): Deleted.
(WebCore::Pasteboard::createForCopyAndPaste): Deleted.
(WebCore::Pasteboard::createForDragAndDrop): Deleted.
(WebCore::Pasteboard::create): Deleted.
(WebCore::Pasteboard::Pasteboard): Deleted.
(WebCore::Pasteboard::selectionData const): Deleted.
(WebCore::selectionDataTypeFromHTMLClipboardType): Deleted.
(WebCore::Pasteboard::writeString): Deleted.
(WebCore::Pasteboard::writePlainText): Deleted.
(WebCore::Pasteboard::write): Deleted.
(WebCore::Pasteboard::writeTrustworthyWebURLsPboardType): Deleted.
(WebCore::Pasteboard::clear): Deleted.
(WebCore::Pasteboard::setDragImage): Deleted.
(WebCore::Pasteboard::hasData): Deleted.
(WebCore::Pasteboard::typesSafeForBindings): Deleted.
(WebCore::Pasteboard::typesForLegacyUnsafeBindings): Deleted.
(WebCore::Pasteboard::readOrigin): Deleted.
(WebCore::Pasteboard::readString): Deleted.
(WebCore::Pasteboard::readStringInCustomData): Deleted.
(WebCore::Pasteboard::fileContentState): Deleted.
(WebCore::Pasteboard::writeMarkup): Deleted.
(WebCore::Pasteboard::writeCustomData): Deleted.
(WebCore::Pasteboard::changeCount const): Deleted.
* Source/WebCore/platform/wpe/PasteboardWPE.cpp:
(WebCore::Pasteboard::createForCopyAndPaste): Deleted.
(WebCore::Pasteboard::Pasteboard): Deleted.
(WebCore::Pasteboard::writeString): Deleted.
(WebCore::Pasteboard::writePlainText): Deleted.
(WebCore::Pasteboard::write): Deleted.
(WebCore::Pasteboard::writeTrustworthyWebURLsPboardType): Deleted.
(WebCore::Pasteboard::clear): Deleted.
(WebCore::Pasteboard::hasData): Deleted.
(WebCore::Pasteboard::typesSafeForBindings): Deleted.
(WebCore::Pasteboard::typesForLegacyUnsafeBindings): Deleted.
(WebCore::Pasteboard::readOrigin): Deleted.
(WebCore::Pasteboard::readString): Deleted.
(WebCore::Pasteboard::readStringInCustomData): Deleted.
(WebCore::Pasteboard::fileContentState): Deleted.
(WebCore::Pasteboard::writeMarkup): Deleted.
(WebCore::Pasteboard::writeCustomData): Deleted.
(WebCore::Pasteboard::changeCount const): Deleted.
* Source/WebKit/SourcesWPE.txt:
* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::startDrag):
* Source/WebKit/UIProcess/WebPageProxy.h:
* Source/WebKit/UIProcess/WebPageProxy.messages.in:
* Source/WebKit/WebProcess/WebCoreSupport/WebDragClient.cpp:
* Source/WebKit/WebProcess/WebCoreSupport/wpe/WebDragClientWPE.cpp: Copied from Source/WebKit/WebProcess/WebCoreSupport/WebDragClient.cpp.
Basic implementation copied from GTK.

(WebKit::WebDragClient::didConcludeEditDrag):
(WebKit::WebDragClient::startDrag):
